Reiniciando feature, iniciando a recriação do DetalheDasPrestacoes/index.js
