+++ src/services/escolas/PrestacaoDeContas.service.js
@@ -55,18 +55,12 @@ export const getDespesasPrestacaoDeContas = async (
   ).data
 }
 
-export const getReceitasPrestacaoDeContas = async (
-  uuidPrestacaoDeContas,
-  acao_associacao__uuid,
-  conferido
-) => {
-  return (
-    await api.get(
-      `/api/prestacoes-contas/${uuidPrestacaoDeContas}/receitas/?acao_associacao_uuid=${acao_associacao__uuid}&conferido=${conferido}`,
-      authHeader
-    )
-  ).data
-}
+export const getReceitasPrestacaoDeContas = async (uuidPrestacaoDeContas, acao_associacao__uuid, conferido) => {
+
+  //  /api/conciliacoes/receitas/?periodo={uuid}&conta_associacao={uuid}&acao_associacao={uuid}&conferido=
+  
+  return ( await api.get(`/api/prestacoes-contas/${uuidPrestacaoDeContas}/receitas/?acao_associacao_uuid=${acao_associacao__uuid}&conferido=${conferido}`,authHeader)).data
+};
 
 export const getConciliarReceita = async (
   uuid_receita,
